+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.0.3] - 2024-03-29
+
+### Added
+ - Introduce the new class OdooModel to use syntax such:
+```python
+m_order = odoo_cli.model('sale.order')
+order_ids = m_order.search(['partner_id', '=', 5])
+```
+
+ - Add methods:
+      - read
+      - read_group
+      - search_count
+
 ## [1.0.2] - 2024-02-05
 
 ### Added